app/requirements.txt
python-gitlab==4.1.1
